requirements.txt
# requirements.txt
# Минимальный набор для локальных make-команд и утилит:
#   make setup, models, models-ping  — src/utils/check_local.py
#   make download, inspect           — src/download.py
#   make sfdb, sfdb-check, sfmodel   — src/utils/download_*.py
#
# Docker-зависимости экспериментов — в experiments/exp0X/requirements.txt

openai
anthropic
huggingface_hub
python-dotenv
PyYAML
pandas
streamlit
